web-app/requirements.txt
flask
requests
pandas
numpy
matplotlib
scipy
transformers
torch
wordcloud
azure-storage-blob
azure-cosmos
dotenv